[MOL-12849] Rename location input group to location field
