+++ src/components/fields/location-field/index.ts
@@ -1,0 +1,2 @@
+export * from "./location-field";
+export * from "./types";
