+++ src/components/fields/location-field/location-modal/infinite-scroll/loading-indicator/config.ts
