+++ src/components/fields/location-field/static-map/index.ts
