+++ src/components/fields/location-field/static-map/static-map.tsx
