--- src/components/fields/location-input-group/index.ts
@@ -1,2 +1,0 @@
-export * from "./location-input-group";
-export * from "./types";
